add invoke

+++ pkg/vm/loader/classloader/prepare.go
@@ -4,7 +4,7 @@ import "github.com/YEXINGZHE54/myvm/pkg/vm/engine/reflect"
 
 func prepare(c *reflect.Class)  {
 	// calc instance field slots
-	slotId := uint(0)
+	slotId := 0
 	if c.Super != nil && c.Super.InstanceSlotCount > 0 {
 		slotId = c.Super.InstanceSlotCount
 	}
@@ -20,7 +20,7 @@ func prepare(c *reflect.Class)  {
 	}
 	c.InstanceSlotCount = slotId
 	// calc static field slots
-	slotId = uint(0)
+	slotId = 0
 	for _, field := range c.Fields {
 		if field.IsStatic() {
 			//only static vars
